Date: 10 Nov, 2016
changeCase.cpp: Write a program that reads a file and creates a new file with
the same data, except reverse the case on the second file.
Everywhere uppercase letters appear in the first file, write
lowercase letters to the new file, and everywhere lowercase
letters appear in the first file, write uppercase letters to the
new file
+
Write a program that displays the number of nonalphabetic
characters in a file

personalRecord.cpp and personalRecord.h: Construct a class to hold personnel records. Use the following
data members, and keep them private:
char name[25];
float salary;
char date_of_birth[9];
Create two constructors, one to initialize the record with its
necessary values and another to create an uninitialized record.
Create member functions to alter the individuals name, salary, and
date of birth.